Введение компонентного подхода
